Split the Crawler from fetch_data.py
